fixed findbugs warning about boxing/unboxing/boxing again in string format
git-svn-id: svn+ssh://svn.code.sf.net/p/jillion/code/trunk@1271 d1f56cec-0f3b-4699-95c7-aedf87c10f2a

+++ src/org/jcvi/common/core/seq/fastx/fasta/pos/DefaultPositionSequenceFastaRecord.java
@@ -48,7 +48,7 @@ public class DefaultPositionSequenceFastaRecord <T extends Sequence<ShortSymbol>
        Iterator<ShortSymbol> iter = positions.iterator();
        int i=1;
        while(iter.hasNext()){
-    	   result.append(String.format("%04d", iter.next().getValue().shortValue()));
+    	   result.append(String.format("%04d", iter.next().getValue()));
     	   if(iter.hasNext()){
     		   if(i%12 == 0){
                    this.appendCarriageReturnAndLineFeed(result);
